# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NotedThat HTTP API",
    "description": "The versioned machine API of a NotedThat server. Paths are relative to the server URL, `/api/v1`.\n\nThis document is generated from the server's code and is authoritative for paths, parameters, schemas and status codes. `docs/API.md` explains behaviour; where the two disagree, `docs/API.md` is wrong.\n\nEvery response carries `x-request-id`, echoing the request's when it sent one.",
    "contact": {
      "name": "NotedThat Contributors"
    },
    "license": {
      "name": "MPL-2.0",
      "identifier": "MPL-2.0"
    },
    "version": "0.12.3"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/knowledgebases": {
      "get": {
        "tags": [
          "knowledgebases"
        ],
        "summary": "List the knowledge bases the caller may see.",
        "description": "Anonymous callers see those whose access policy grants `anyone` something;\nan anonymous caller who can see none is answered `401`, not an empty list.",
        "operationId": "list_kbs",
        "responses": {
          "200": {
            "description": "The knowledge bases the caller may see.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KbListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "408": {
            "$ref": "#/components/responses/RequestTimeout"
          },
          "503": {
            "$ref": "#/components/responses/BackendUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ]
      }
    },
    "/knowledgebases/{kb_slug}": {
      "get": {
        "tags": [
          "knowledgebases"
        ],
        "summary": "List one page of a knowledge base's objects.",
        "description": "Only the objects the caller's `list` grant spans are returned. A page can\ncome back short of `limit` with `truncated` set; keep paging.",
        "operationId": "list_objects",
        "parameters": [
          {
            "name": "kb_slug",
            "in": "path",
            "description": "The knowledge base.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/KbSlug"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "description": "Only keys starting with this prefix.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size. `0` or absent is `100`; larger than `1000` is `1000`.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 1000,
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "The `next_cursor` of the previous page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "One page of objects.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "408": {
            "$ref": "#/components/responses/RequestTimeout"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/BackendUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {}
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "description": "The error envelope every JSON error response carries.",
        "required": [
          "error",
          "message",
          "request_id"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Machine-readable code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation. Its wording is not part of the contract."
          },
          "request_id": {
            "type": "string",
            "description": "The request's id, the same value as the `x-request-id` response header."
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Machine-readable error code. Clients branch on this, never on `message`.",
        "enum": [
          "invalid_request",
          "malformed_range",
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict",
          "gone",
          "payload_too_large",
          "precondition_failed",
          "no_match",
          "ambiguous_match",
          "precondition_required",
          "internal_error",
          "backend_unavailable",
          "request_timeout"
        ]
      },
...
      "KbListEntry": {
        "type": "object",
        "description": "One knowledge base: the slug the routes take, and the manifest's display\nname and description, so an agent can choose where to search before it\nsearches.",
        "required": [
          "kb_slug",
          "display_name"
        ],
        "properties": {
          "kb_slug": {
            "type": "string",
            "description": "The slug the routes take."
          },
          "display_name": {
            "type": "string",
            "description": "The manifest's display name; the slug when it has none."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "The manifest's description, when it has one."
          }
        }
      },
      "KbListResponse": {
        "type": "object",
        "description": "The knowledge bases the caller may see.",
        "required": [
          "knowledgebases"
        ],
        "properties": {
          "knowledgebases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KbListEntry"
            },
            "description": "In slug order."
          }
        }
      },
      "KbSlug": {
        "type": "string",
        "description": "A knowledge base slug: 1 to 40 of `a-z`, `0-9` and `-`, not starting or ending with `-`.",
        "pattern": "^[a-z0-9]([a-z0-9-]{0,38}[a-z0-9])?$"
      },
...
      "ObjectListResponse": {
        "type": "object",
        "description": "One page of a knowledge base's objects.",
        "required": [
          "objects",
          "truncated"
        ],
        "properties": {
          "objects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectMeta"
            },
            "description": "The objects on this page."
          },
          "truncated": {
            "type": "boolean",
            "description": "Whether more objects follow; fetch them with `next_cursor`."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "The `cursor` of the next page; `null` on the last one."
          }
        }
      },
      "ObjectMeta": {
        "type": "object",
        "description": "Metadata about a single stored object, returned by HEAD and LIST responses.",
        "required": [
          "key",
          "size"
        ],
        "properties": {
          "key": {
            "type": "string",
            "description": "The object's key (path within the KB bucket, without leading slash)."
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "Size in bytes.",
            "minimum": 0
          },
          "last_modified": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Last-modified Unix timestamp (seconds), if the backend provides one."
          },
          "content_type": {
            "type": [
              "string",
              "null"
            ],
            "description": "Content-Type as stored (echoed from the original PUT)."
          },
          "etag": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque `ETag` from the backend, wrapped in quotes per RFC 7232 §2.3\n(e.g., `\"\\\"abc123\\\"\"`). Emitted verbatim in HTTP responses."
          }
        }
      },
...
      "Refusal": {
        "type": "object",
        "description": "The error envelope of a request the server refused before handling it: at its limit of requests in flight, or out of time.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "`backend_unavailable` or `request_timeout`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation. Its wording is not part of the contract."
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The request's id, when one had been assigned."
          }
        }
      },
...
    },
    "responses": {
      "BackendUnavailable": {
        "description": "`backend_unavailable`: a backend or queue is unavailable or full, or the server is at its limit of requests in flight. Retry after `Retry-After`. A write answered this way may already be stored; repeat it.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "string"
            },
            "description": "Seconds to wait: `5`."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Error"
                },
                {
                  "$ref": "#/components/schemas/Refusal"
                }
              ]
            }
          }
        }
      },
      "BadRequest": {
        "description": "The request is malformed: `invalid_request` or `malformed_range`. A query string the server cannot parse at all is answered `400` with a plain-text body instead.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "`forbidden`: the credential is valid and the knowledge base's access policy does not grant this operation. An anonymous caller is answered `404` instead.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "GatewayTimeout": {
        "description": "`request_timeout`: the request did not complete within the server's timeout.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Refusal"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "`internal_error`: the server failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "`not_found`: no such knowledge base or object — or one an anonymous caller may not see, which reads identically.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "RequestTimeout": {
        "description": "`request_timeout`: the request body stopped arriving for longer than the server's idle limit.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Refusal"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "`unauthorized`: no credential where one is required, or one that did not verify.",
        "headers": {
          "WWW-Authenticate": {
            "schema": {
              "type": "string"
            },
            "description": "`Bearer resource_metadata=\"…\"`, when the deployment publishes OAuth protected resource metadata."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "service token or OIDC access token (JWT)",
        "description": "`Authorization: Bearer <token>`: the deployment's service token, or an access token from the configured OIDC issuer. A malformed or unverified credential is always `401`, never treated as anonymous."
      }
    }
  },
...
}